+++ doc/changelog.rst
@@ -9,11 +9,14 @@ General
 ^^^^^^^
 
 - Added primitive collider for sphere-cylinder contacts, previously this pair used the generic convex-convex collider.
+- Added an optional ``content_type`` attribute to hfield, texture, and mesh assets. This attribute supports a formatted
+  `MIME <https://en.wikipedia.org/wiki/MIME>`_ string used to determine the type of the asset file without resorting to
+  pulling the type from the file extension.
 - Added analytic derivatives for quaternion :ref:`subtraction<mjd_subQuat>` and :ref:`integration<mjd_quatIntegrate>`
   (rotation with an angular velocity). Derivatives are in the 3D tangent space.
 - Added :ref:`mjv_connector` which has identical functionality to :ref:`mjv_makeConnector`, but with more convenient
   "from-to" argument parametrization. :ref:`mjv_makeConnector` is now deprecated.
-- Bump oldest supported MacOS from version 10.12 to 11. MacOS 11 is the oldest version still maintained by Apple.
+- Bumped oldest supported MacOS from version 10.12 to 11. MacOS 11 is the oldest version still maintained by Apple.
 
 Python bindings
 ^^^^^^^^^^^^^^^